V2 extended transactions (#287)
Read/write and readonly transactions use completely different concepts in client library (`TransactionManager` is only for read/write transactions), so I created `DatabaseClientTransactionManager` to encapsulate all the random bits of transactional state.

All forms of stale readonly transactions are supported with this PR.

Partitioned DML is out of scope, since it does not fit into beginTransaction-runQueries-commitTransaction pattern, and it's not atomic anyway. It also does not happen to be supported asynchronously in the client library (although it is in Connection API, but more as a wrap-this-in-a-thread workaround). 

Fixes #249 .