Upgrade SPI and fix blocking errors (#489)
I've disabled the TCK tests and filed a tracking issue #488 for it. There are still changes in `SpannerClientLibraryTestKit.java` for easy-to-fix failures due to test column being renamed from "value" to "test_value".

The remaining TCK tests are for missing features, but will need to be re-enabled before we can release (for assurance on TCK compliance).